Introduce promote.hpp for numpy-like type promotion rules
