+++ mdtensor/core/broadcast/bcast_to.hpp
@@ -46,8 +46,7 @@ namespace mdtensor::core {
     } else {
         using index_t = typename exts_t::index_type;
         using cindex_t =
-            core::common_integral_type_t<typename in_mds_t::index_type,
-                                         index_t>;
+            std::common_type_t<typename in_mds_t::index_type, index_t>;
 
         // ni = new_rank - org_rank + oi
         const auto get_ni = [](std::size_t i) {
