+++ mdtensor/logic/isclose.hpp
@@ -42,9 +42,8 @@ constexpr void isclose_ufunc(auto &&in1, auto &&in2, auto &&out, auto &&rtol,
         }
     }
 
-    using common_t =
-        core::common_arithmetic_type_t<decltype(in1), decltype(in2),
-                                       decltype(rtol), decltype(atol)>;
+    using common_t = core::promote_type_t<decltype(in1), decltype(in2),
+                                          decltype(rtol), decltype(atol)>;
 
     out = absolute(static_cast<common_t>(in1) - static_cast<common_t>(in2)) <=
           (static_cast<common_t>(atol) +
