--- tests/core/base/type/common_integral/BUILD.bazel
@@ -1,19 +1,0 @@
-load("@rules_cc//cc:defs.bzl", "cc_test")
-
-cc_test(
-    name = "main",
-    srcs = [
-        "main.cpp",
-        ],
-    copts = [
-        "-O3",
-        "-funroll-loops",
-        "-fno-rtti",
-        "-Wall",
-        "-march=native",
-    ],
-    deps = [
-        "@googletest//:gtest_main",
-        "//:mdtensor",
-    ],
-)
